Refactor tests, add new tests
